moving database roation code to fattest.simplicity
