+++ dev/com.ibm.ws.concurrent.persistent_fat/fat/src/com/ibm/ws/concurrent/persistent/fat/PersistentExecutorWithFailoverEnabledTest.java
@@ -25,7 +25,9 @@ import com.ibm.websphere.simplicity.config.ServerConfiguration;
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
 import componenttest.custom.junit.runner.FATRunner;
-import componenttest.topology.database.DerbyEmbeddedUtilities;
+import componenttest.topology.database.container.DatabaseContainerFactory;
+import componenttest.topology.database.container.DatabaseContainerType;
+import componenttest.topology.database.container.DatabaseContainerUtil;
 import componenttest.topology.impl.LibertyFileManager;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
@@ -36,7 +38,7 @@ import web.SchedulerFATServlet;
 public class PersistentExecutorWithFailoverEnabledTest extends FATServletClient {
 
     private static final String APP_NAME = "schedtest";
-    private static final String DB_NAME = "${shared.resource.dir}/data/scheddb";
+    private static final String DB_NAME = APP_NAME;
 
     private static ServerConfiguration originalConfig;
 
@@ -57,7 +59,6 @@ public class PersistentExecutorWithFailoverEnabledTest extends FATServletClient 
         // Delete the Derby database that might be used by the persistent scheduled executor and the Derby-only test database
         Machine machine = server.getMachine();
         String installRoot = server.getInstallRoot();
-        LibertyFileManager.deleteLibertyDirectoryAndContents(machine, installRoot + "/usr/shared/resources/data/scheddb");
         LibertyFileManager.deleteLibertyDirectoryAndContents(machine, installRoot + "/usr/shared/resources/data/testdb");
 
         originalConfig = server.getServerConfiguration();
