+++ dev/com.ibm.ws.concurrent.persistent_fat_timers/bnd.bnd
@@ -34,9 +34,5 @@ fat.test.databases: true
 	org.testcontainers:database-commons;version=1.12.3,\
 	org.testcontainers:jdbc;version=1.12.3,\
 	org.testcontainers:testcontainers;version=1.12.3,\
-	org.testcontainers:db2;version=1.12.3,\
-	org.testcontainers:postgresql;version=1.12.3,\
-	org.testcontainers:mssqlserver;version=1.12.3,\
-	org.testcontainers:oracle-xe;version=1.12.3,\
 	org.rnorth.duct-tape:duct-tape;version=1.0.7,\
 	org.slf4j:slf4j-api;version=1.7.7
