+++ dev/fattest.simplicity/bnd.bnd
@@ -11,6 +11,9 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
+javac.source: 1.8
+javac.target: 1.8
+
 Bundle-Name: FAT infrastructure
 Bundle-SymbolicName: fattest.simplicity
 Bundle-Description: FAT infrastructure; version=${bVersion}
@@ -38,6 +41,7 @@ Export-Package: \
 	componenttest.rules;version=1.0.16,\
 	componenttest.rules.repeater;version=1.0.16,\
 	componenttest.topology.database;version=1.0.16,\
+	componenttest.topology.database.container;version=1.0.16,\
 	componenttest.topology.impl;version=1.0.16,\
 	componenttest.topology.ldap;version=1.0.16,\
 	componenttest.topology.utils;version=1.0.16,\
@@ -84,4 +88,13 @@ test.project: true
 	org.jboss.arquillian.core:arquillian-core-spi;version=1.3.0.Final,\
 	org.jboss.arquillian.core:arquillian-core-api;version=1.3.0.Final,\
     org.jboss.arquillian.test:arquillian-test-spi;version=1.3.0.Final,\
-    org.testcontainers:testcontainers;version=1.12.3
+    org.testcontainers:testcontainers;version=1.12.3,\
+    org.testcontainers:database-commons;version=1.12.3,\
+	org.testcontainers:jdbc;version=1.12.3,\
+	org.testcontainers:testcontainers;version=1.12.3,\
+	org.testcontainers:db2;version=1.12.3,\
+	org.testcontainers:postgresql;version=1.12.3,\
+	org.testcontainers:mssqlserver;version=1.12.3,\
+	org.testcontainers:oracle-xe;version=1.12.3,\
+	org.rnorth.duct-tape:duct-tape;version=1.0.7,\
+	org.slf4j:slf4j-api;version=1.7.7
