+++ dev/fattest.simplicity/src/componenttest/topology/database/container/DatabaseContainerType.java
@@ -8,10 +8,10 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.concurrent.persistent.fat.timers;
+package componenttest.topology.database.container;
 
-import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.Db2Container;
+import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.MSSQLServerContainer;
 import org.testcontainers.containers.OracleContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
@@ -20,9 +20,7 @@ import org.testcontainers.containers.output.OutputFrame;
 import com.ibm.websphere.simplicity.log.Log;
 
 /**
- * This is a current list of database test-containers that are in the database rotation.
- *
- * TODO: Since this enum is general it should be moved to fattest simplicity in the future
+ * This enum represents the current list of database testcontainers that are in the database rotation.
  */
 @SuppressWarnings("rawtypes")
 public enum DatabaseContainerType {
@@ -32,9 +30,9 @@ public enum DatabaseContainerType {
     Postgre("postgresql.jar", PostgreSQLContainer.class),
     SQLServer("mssql-jdbc.jar", MSSQLServerContainer.class);
 
-	private String driverName;
-	private Class<? extends JdbcDatabaseContainer> clazz;
-	
+    private String driverName;
+    private Class<? extends JdbcDatabaseContainer> clazz;
+
     DatabaseContainerType(final String driverName, final Class<? extends JdbcDatabaseContainer> clazz) {
         this.driverName = driverName;
         this.clazz = clazz;
@@ -43,37 +41,40 @@ public enum DatabaseContainerType {
     /**
      * Returns the common JDBC Driver name for this test-container type.
      * Example: 'ojdbc8_g.jar'
+     * 
      * @return String - JDBC Driver Name
      */
     public String getDriverName() {
         return driverName;
     }
-    
+
     /**
      * Returns test-container class associated with this test-container type.
+     * 
      * @return Java Class
      */
     public Class getContainerClass() {
-    	return clazz;
+        return clazz;
     }
-    
+
     /**
      * Given a JDBC test-container return the corresponding Database Container Type.
-     * @param cont - A database container. 
-     * @return DatabaseContainerType - type enum
+     * 
+     * @param  cont - A database container.
+     * @return      DatabaseContainerType - type enum
      */
     public static DatabaseContainerType valueOf(JdbcDatabaseContainer cont) {
-    	for(DatabaseContainerType elem : values())
-    		if(elem.getContainerClass() == cont.getClass())
-    			return elem;
-    	throw new IllegalArgumentException("Unrecognized JdbcDatabaseContainer class: " + cont.getClass().getCanonicalName());
+        for (DatabaseContainerType elem : values())
+            if (elem.getContainerClass() == cont.getClass())
+                return elem;
+        throw new IllegalArgumentException("Unrecognized JdbcDatabaseContainer class: " + cont.getClass().getCanonicalName());
     }
-    
-	//Private Method: used to setup logging for containers to this class. 
-	public void log(OutputFrame frame) {
+
+    //Private Method: used to setup logging for containers to this class. 
+    public void log(OutputFrame frame) {
         String msg = frame.getUtf8String();
         if (msg.endsWith("\n"))
             msg = msg.substring(0, msg.length() - 1);
         Log.info(this.clazz, "output", msg);
-	}
+    }
 }
